split telemetry constant on new file
