+++ src/extension/src/constants/constants.ts
@@ -76,24 +76,6 @@ export const CONSTANTS = {
 
 export const PROJECT_NAME_VALIDATION_LIMIT = 50;
 
-export enum TelemetryEventName {
-  ExtensionLaunch = "Extension-Launch",
-  ExtensionClosed = "Extension-closed",
-  WizardSession = "Wizard-To-Generate-Session-Time",
-  AppServiceDeploy = "Azure-App-Service-Deployment",
-  CosmosDBDeploy = "Azure-Cosmos-Deployment",
-  ResourceGroupDeploy = "Azure-Resource-Group-Deployment",
-  PageChange = "Wizard-Page-Change",
-  CreateNewProject = "Create-New-Project",
-  SyncEngine = "Sync-Engine",
-  ConnectionStringReplace = "Connection-String-Replaced",
-  TrackOpenAddPagesModal = "Open-Add-Pages-Modal",
-  TrackPressQuickstart = "Press-Quickstart",
-  OpenAppServiceModalFromServicesList = "Open-AppService-Modal-From-Services-List",
-  OpenCosmosDBServiceModalFromServicesList = "Open-CosmosDBService-Modal-From-Services-List",
-  OpenAzureServiceAdvancedMode = "Open-Azure-Service-Advanced-Mode"
-}
-
 export const PAYLOAD_MESSAGES_TEXT = {
   RESET_PAGES_TEXT: "Sending reset pages request...",
   SWITCH_FRAMEWORKS_TEXT: "Sending framework change request...",
